requirements.txt
# AWS CDK
aws-cdk-lib==2.95.1
constructs>=10.0.0,<11.0.0

# AWS SDK
boto3==1.28.57
botocore==1.31.57

# Lambda runtime dependencies
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0

# Development dependencies
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
black==23.9.1
flake8==6.1.0
mypy==1.5.1
isort==5.12.0

# For local development
python-dotenv==1.0.0
httpx==0.24.1
